third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-basic.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#e {
    --foo: {
        background-color: green;
        width: 100px;
        height: 200px;
    };
    @apply --foo;
}
</style>
<div id=e>
</div>
<script>
test(function(){
    assert_equals(getComputedStyle(e).backgroundColor, "rgb(0, 128, 0)");
    assert_equals(getComputedStyle(e).width, "100px");
    assert_equals(getComputedStyle(e).height, "200px");
}, "Basic usage of @apply with non-shorthand properties");
</script>
